Memory Optimization for cub_sort_key_value
cub_sort_key_value uses the cub::DeviceRadixSort::SortPairs function. The current use of SortPairs requires an additional memory of O(N+P) where N is length of input and P is number of streaming multiprocessors on the device.

Modifying the cub_sort_key_value function to use the DoubleBuffer overload of cub::DeviceRadixSort::SortPairs allows us to
perform the same operation in O(P) time. The reduced memory requirement is because the sorting is done inplace. As per the documentation there is no performance degradation due to using this overload.

Additionally, the sorted result (keys and values) is copied from the current buffer to the alternate buffer using memcpy_kernel.